# solution/MainServer/swagger/swaggerDocumentation.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IumTweb_project",
    "description": "This API is part of the IUM-TWEB project for the 2023-2024 academic year at the Dipartimento di Informatica, Università di Torino. Designed to enable both fans and pundits to access and analyze a wide range of football data, this API supports functionalities across multiple servers (two Express servers and one Java Spring server). It provides interfaces for querying football competitions, clubs, and player statistics, and integrates with MongoDB and Postgres databases to manage dynamic and static datasets respectively. [ IUM-TWEB Group Accornero, Aiello, Bonincontro ]",
    "version": "1.0.0",
    "contact": {
      "name": "IUM-TWEB Group Accornero, Aiello, Bonincontro"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "MainServer - Can handle all API requests."
    },
    {
      "url": "http://localhost:3001",
      "description": "ExpressServer"
    },
    {
      "url": "http://localhost:8081",
      "description": "SpringServer"
    }
  ],
  "paths": {
...
    "/players/getAllPlayers": {
      "get": {
        "tags": [
          "Players"
        ],
        "summary": "Retrieve all players",
        "description": "Retrieves a list of all players in the database.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "MainServer"
          },
          {
            "url": "http://localhost:8081",
            "description": "SpringServer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful - A list of all players in the database.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An error occurred on the server while processing the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/players/searchByClubIdAndSeason": {
      "get": {
        "tags": [
          "Players"
        ],
        "summary": "Search players by club ID and season",
        "description": "Retrieves a list of players who played for a given club in a specified season.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "MainServer"
          },
          {
            "url": "http://localhost:8081",
            "description": "SpringServer"
          }
        ],
        "parameters": [
          {
            "name": "club_id",
            "in": "query",
            "required": true,
            "description": "The unique identifier for the club",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": true,
            "description": "The season for which to retrieve player data",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful - A list of all players that played in a certain club in a certain season",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An error occurred on the server while processing the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/players/getPlayersByCompetition/{competition_id}": {
      "get": {
        "tags": [
          "Players"
        ],
        "summary": "Get player by competition ID",
        "description": "Retrieves players based on a competition ID.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "MainServer"
          },
          {
            "url": "http://localhost:8081",
            "description": "SpringServer"
          }
        ],
        "parameters": [
          {
            "name": "competition_id",
            "in": "path",
            "required": true,
            "description": "The unique identifier for the competition",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful - A list of players that are playing in the competition specified by competition_id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An error occurred on the server while processing the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/players/getPlayersByCompetitionAndLastSeasonSortedByValue/{competition_id}/{last_season}": {
      "get": {
        "tags": [
          "Players"
        ],
        "summary": "Get player by competition ID and last season, sorted by value",
        "description": "Retrieves players based on a competition ID and the last season, sorted by market value in a descending manner.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "MainServer"
          },
          {
            "url": "http://localhost:8081",
            "description": "SpringServer"
          }
        ],
        "parameters": [
          {
            "name": "competitionId",
            "in": "path",
            "required": true,
            "description": "The competition's unique identifier",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lastSeason",
            "in": "path",
            "required": true,
            "description": "The last season of interest",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful - A list of players that played in a certain competition and in a certain season, sorted by value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An error occurred on the server while processing the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/players/getAllDomesticCompetitions": {
      "get": {
        "tags": [
          "Players"
        ],
        "summary": "Get information about domestic competition",
        "description": "Returns all the current domestic competition basic info(id and name) that are in the Players table, sorted alphabetically by competition name.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "MainServer"
          },
          {
            "url": "http://localhost:8081",
            "description": "SpringServer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful - A list of small competition object with two fields: id and name.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example":"IT1"
                      },
                      "name": {
                        "type": "string",
                        "example": "serie-a"
                      }
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An error occurred on the server while processing the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/players/getAllCountryOfCitizenship": {
      "get": {
        "tags": [
          "Players"
        ],
        "summary": "Get all the nationalities",
        "description": "Retrives all the country of citizenship that are in the Players table, sorted alphabetically.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "MainServer"
          },
          {
            "url": "http://localhost:8081",
            "description": "SpringServer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful - A list of strings that are all the nationalities (Without null).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An error occurred on the server while processing the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/players/getTop150PlayersByMarketValue": {
      "get": {
        "tags": [
          "Players"
        ],
        "summary": "Get top 150 players by market value",
        "description": "Retrieves a list of the top 150 players sorted by their market value in a descending manner.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "MainServer"
          },
          {
            "url": "http://localhost:8081",
            "description": "SpringServer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful - A list of top 150 players.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An error occurred on the server while processing the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/players/findPlayersByLetterInName": {
      "get": {
        "tags": [
          "Players"
        ],
        "summary": "Find players by letter in name",
        "description": "Retrieves a list of players whose names contain a specified letter.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "MainServer"
          },
          {
            "url": "http://localhost:8081",
            "description": "SpringServer"
          }
        ],
        "parameters": [
          {
            "name": "letter",
            "in": "query",
            "required": true,
            "description": "The letter to search for in player names",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful - A list of players.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An error occurred on the server while processing the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/players/getPlayersImgUrlById": {
      "get": {
        "tags": ["Players"],
        "summary": "Get players' image URLs by player ID",
        "description": "Returns image URLs associated with player IDs.",
        "servers": [
          {
            "url": "http://localhost:3000",
            "description": "MainServer"
          },
          {
            "url": "http://localhost:8081",
            "description": "SpringServer"
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "starting",
            "schema": {
              "type": "array",
              "items":{
                "type": "integer"
              }
            },
            "description": "A list of integers, contains all the player IDs of the players who start a match"
          },
          {
            "in": "query",
            "name": "substitutes",
            "schema": {
              "type": "array",
              "items":{
                "type": "integer"
              }
            },
            "description": "A list of integers, contains all the player IDs of the players who are on the bench"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful - Returns image URLs of players associated with player_id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "playerId": {
                        "type": "integer"
                      },
                      "imageUrl": {
                        "format": "url",
                        "description": "URL of the player's image"
                      }
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An error occurred on the server while processing the request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": " :( - Error Message"
          },
          "internal_code": {
            "type": "string",
            "description": " :( - Code Error"
          }
        }
      }
    }
  }
}